backend/requirements.txt
fastapi
uvicorn[standard]
chromadb
google-genai
python-dotenv
playwright
groq
